copy: don't move on remote if a copy has already been made
This allows us to move the stuff to download out of the way on the
remote, then continue scraping while we download.

+++ Makefile
@@ -15,7 +15,7 @@ copyDbFromRemote:
 	ssh "$(SSH_HOST)" bash << HERE
 		cd /home/kisaragi/goo.gl-archive/
 		make checkpoint
-		mv data.sqlite download.sqlite
+		[ -f download.sqlite ] || mv data.sqlite download.sqlite
 	HERE
 	rsync -h -P -z "$(SSH_HOST):/home/kisaragi/goo.gl-archive/download.sqlite" remote-data.sqlite
